WiFi: Enable reconnect and add bootmanager
Reconnects when there is an existing configuration in the WifiControl
plugin. Removes configs when a connect fails for a failsafe.

Also enable config.json to redirect to bootmanager by operator or by url
value. If config.json is not valid or bootmanager call fails the app
will go to the default RDK boot manager.